requirements.txt
absl-py==1.0.0
asn1crypto==0.24.0
astunparse==1.6.3
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
chardet==3.0.4
charset-normalizer==2.0.12
cloud-init==19.1.16
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
contextlib2==21.6.0
coverage==6.2
cryptography==2.1.4
cycler==0.11.0
dataclasses==0.8
decorator==4.4.2
distro-info===0.18ubuntu0.18.04.1
Django==2.2.25
future==0.18.2
gast==0.3.3
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.45.0
h5py==2.10.0
hyperopt==0.1.2
idna==2.8
importlib-metadata==1.7.0
importlib-resources==3.0.0
Jinja2==2.11.2
joblib==0.14.1
json-tricks==3.13.2
jsonpatch==1.32
jsonpointer==2.2
jsonschema==4.0.0
Keras==2.1.6
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.1.0
language-selector==0.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
netifaces==0.10.4
networkx==2.3
numpy==1.18.4
nvidia-ml-py3==7.352.0
oauthlib==3.2.0
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.4
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycrypto==2.6.1
PyGObject==3.26.1
pymongo==4.1.1
pyparsing==3.0.8
pyrsistent==0.18.0
python-apt==1.6.5+ubuntu0.7
python-dateutil==2.8.2
pytz==2022.1
pyxdg==0.25
PyYAML==6.0
pyzmq==22.3.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
schema==0.7.5
scikit-learn==0.24.2
scipy==1.4.1
SecretStorage==2.3.1
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
ssh-import-id==5.7
tensorboard==2.2.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.2.0
tensorflow-estimator==2.2.0
termcolor==1.1.0
threadpoolctl==3.1.0
torch==1.10.1
torchvision==0.11.2
tqdm==4.36.1
typing_extensions==4.1.1
ubuntu-advantage-tools==27.6
ufw==0.36
urllib3==1.26.9
Werkzeug==2.0.3
wrapt==1.14.0
zipp==3.6.0
zmq==0.0.0
